# Use w-NAF in `ka_orchard`