--- a/Cargo.toml
+++ b/Cargo.toml
@@ -29,7 +29,7 @@ fpe = "0.4"
 group = "0.10"
 halo2 = "0.0"
 lazy_static = "1"
-pasta_curves = "0.1"
+pasta_curves = "0.1.2"
 proptest = { version = "1.0.0", optional = true }
 rand = "0.8"
 nonempty = "0.7"