Improves the base-line cost of trial decryption by over 40%.