@str4d 2021-08-06
Before:
```
note-decryption/valid   time:   [3.0692 ms 3.0790 ms 3.0895 ms]                                   
                        change: [-0.7686% -0.0930% +0.5480%] (p = 0.78 > 0.05)
                        No change in performance detected.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
note-decryption/invalid time:   [163.46 us 164.20 us 165.15 us]                                    
                        change: [-0.9767% -0.2059% +0.5491%] (p = 0.62 > 0.05)
                        No change in performance detected.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) high mild
  5 (5.00%) high severe
note-decryption/compact-valid                                                                             
                        time:   [3.0680 ms 3.0777 ms 3.0884 ms]
                        change: [-0.2551% +0.3009% +0.8282%] (p = 0.29 > 0.05)
                        No change in performance detected.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
compact-note-decryption/invalid                                                                            
                        time:   [179.00 ms 179.35 ms 179.73 ms]
                        thrpt:  [5.5639 Kelem/s 5.5758 Kelem/s 5.5867 Kelem/s]
                 change:
                        time:   [-3.4445% -3.0790% -2.7340%] (p = 0.00 < 0.05)
                        thrpt:  [+2.8109% +3.1768% +3.5674%]
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
```
After:
```
note-decryption/valid   time:   [2.8425 ms 2.8573 ms 2.8748 ms]                                   
                        change: [-7.7656% -7.2003% -6.5631%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
note-decryption/invalid time:   [87.242 us 87.627 us 88.114 us]                                    
                        change: [-47.103% -46.695% -46.322%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
  3 (3.00%) high severe
note-decryption/compact-valid                                                                             
                        time:   [2.8296 ms 2.8391 ms 2.8502 ms]
                        change: [-8.2200% -7.7522% -7.3151%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
compact-note-decryption/invalid                                                                            
                        time:   [100.23 ms 100.61 ms 101.02 ms]
                        thrpt:  [9.8993 Kelem/s 9.9397 Kelem/s 9.9770 Kelem/s]
                 change:
                        time:   [-44.146% -43.904% -43.660%] (p = 0.00 < 0.05)
                        thrpt:  [+77.493% +78.266% +79.039%]
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
```